fix - minor bug fixed , added a loader component for loading time

+++ src/components/post-form/postForm.jsx
@@ -20,26 +20,24 @@ export default function PostForm({ post }) {
   const userData = useSelector((state) => state.auth.userData);
 
   const submit = async (data) => {
-    if (post) {
-      const file = data.image[0]
-        ? await appwriteService.uploadFile(data.image[0])
-        : null;
-
+    if (post && post.$id && post.featuredimage) {
+      const file = data.image[0] ? await appwriteService.uploadFile(data.image[0]) : null;
+  
       if (file) {
         appwriteService.deleteFile(post.featuredimage);
       }
-
+  
       const dbPost = await appwriteService.updatePost(post.$id, {
         ...data,
         featuredimage: file ? file.$id : undefined,
       });
-
+  
       if (dbPost) {
         navigate(`/post/${dbPost.$id}`);
       }
     } else {
       const file = await appwriteService.uploadFile(data.image[0]);
-
+  
       if (file) {
         const fileId = file.$id;
         data.featuredimage = fileId;
@@ -47,14 +45,14 @@ export default function PostForm({ post }) {
           ...data,
           userid: userData.$id,
         });
-
+  
         if (dbPost) {
           navigate(`/post/${dbPost.$id}`);
         }
       }
     }
   };
-
+  
   const slugTransform = useCallback((value) => {
     if (value && typeof value === "string")
       return value
